[backend] resolve dependencies cycles (#3318)
