+++ opencti-platform/opencti-graphql/src/database/stix-converter.ts
@@ -3,7 +3,7 @@ import { version as uuidVersion } from 'uuid';
 import { extractEntityRepresentative, isEmptyField, isInferredIndex } from './utils';
 import { FunctionalError, UnsupportedError } from '../config/errors';
 import { isBasicObject } from '../schema/stixCoreObject';
-import { isBasicRelationship, isStixRelationship } from '../schema/stixRelationship';
+import { isBasicRelationship } from '../schema/stixRelationship';
 import {
   INPUT_BCC,
   INPUT_BELONGS_TO,
@@ -114,16 +114,11 @@ import {
   INPUT_MARKINGS,
   INPUT_OBJECTS
 } from '../schema/general';
-import { FROM_START, FROM_START_STR, hashValue, UNTIL_END, UNTIL_END_STR } from '../utils/format';
+import { FROM_START, FROM_START_STR, UNTIL_END, UNTIL_END_STR } from '../utils/format';
 import { isRelationBuiltin } from './stix';
 import { isInternalRelationship } from '../schema/internalRelationship';
 import { isInternalObject } from '../schema/internalObject';
 import { ENTITY_TYPE_VOCABULARY } from '../modules/vocabulary/vocabulary-types';
-import { SYSTEM_USER } from '../utils/access';
-import { findById as findStixCoreObject } from '../domain/stixCoreObject';
-import { findById as findUser } from '../domain/user';
-import type { NotificationUser } from '../manager/notificationManager';
-import type { AuthContext } from '../types/user';
 
 export const isTrustedStixId = (stixId: string): boolean => {
   const segments = stixId.split('--');
@@ -1536,208 +1531,6 @@ export const registerStixRepresentativeConverter = (type: string, convertFn: Rep
   stixRepresentativeConverters.set(type, convertFn);
 };
 
-export const extractStixRepresentative = async (context: AuthContext, user: NotificationUser, stix: S.StixObject): Promise<string> => {
-  const entityType = stix.extensions[STIX_EXT_OCTI].type;
-  // region Modules
-  const convertFn = stixRepresentativeConverters.get(entityType);
-  if (convertFn) {
-    return convertFn(stix);
-  }
-  // endregion
-  // region Sighting
-  if (isStixSightingRelationship(entityType)) {
-    const authUser = await findUser(context, SYSTEM_USER, user.user_id);
-    const sighting = stix as SRO.StixSighting;
-    const fromId = sighting.extensions[STIX_EXT_OCTI].sighting_of_ref;
-    const toIds = sighting.extensions[STIX_EXT_OCTI].where_sighted_refs;
-    const fromInstance = await findStixCoreObject(context, authUser, fromId);
-    const toInstances = await Promise.all(toIds.map((toId) => findStixCoreObject(context, authUser, toId)));
-    const fromValue = fromInstance ? sighting.extensions[STIX_EXT_OCTI].sighting_of_value : 'Restricted';
-    const allTargetValues = sighting.extensions[STIX_EXT_OCTI].where_sighted_values;
-    const targetValues = [];
-    for (let index = 0; index < toInstances.length; index += 1) {
-      targetValues.push(toInstances[index] ? allTargetValues[index] : 'Restricted');
-    }
-    return `${fromValue} sighted in/at ${targetValues.join(', ')}`;
-  }
-  // endregion
-  // region Relationship
-  if (isStixRelationship(entityType)) {
-    const authUser = await findUser(context, SYSTEM_USER, user.user_id);
-    const relation = stix as SRO.StixRelation;
-    const fromInstance = await findStixCoreObject(context, authUser, relation.extensions[STIX_EXT_OCTI].source_ref);
-    const toInstance = await findStixCoreObject(context, authUser, relation.extensions[STIX_EXT_OCTI].target_ref);
-    const fromValue = fromInstance ? relation.extensions[STIX_EXT_OCTI].source_value : 'Restricted';
-    const targetValue = toInstance ? relation.extensions[STIX_EXT_OCTI].target_value : 'Restricted';
-    return `${fromValue} ${relation.relationship_type} ${targetValue}`;
-  }
-  // endregion
-  // region Entities
-  if (isStixDomainObjectIdentity(entityType)) {
-    return (stix as SDO.StixIdentity).name;
-  }
-  if (isStixDomainObjectLocation(entityType)) {
-    return (stix as SDO.StixLocation).name;
-  }
-  if (entityType === ENTITY_TYPE_CONTAINER_REPORT) {
-    return (stix as SDO.StixReport).name;
-  }
-  if (entityType === ENTITY_TYPE_MALWARE) {
-    return (stix as SDO.StixMalware).name;
-  }
-  if (entityType === ENTITY_TYPE_INFRASTRUCTURE) {
-    return (stix as SDO.StixInfrastructure).name;
-  }
-  if (entityType === ENTITY_TYPE_ATTACK_PATTERN) {
-    return (stix as SDO.StixAttackPattern).name;
-  }
-  if (entityType === ENTITY_TYPE_CAMPAIGN) {
-    return (stix as SDO.StixCampaign).name;
-  }
-  if (entityType === ENTITY_TYPE_THREAT_ACTOR) {
-    return (stix as SDO.StixThreatActor).name;
-  }
-  if (entityType === ENTITY_TYPE_CONTAINER_NOTE) {
-    return (stix as SDO.StixNote).abstract;
-  }
-  if (entityType === ENTITY_TYPE_CONTAINER_OPINION) {
-    return (stix as SDO.StixOpinion).opinion;
-  }
-  if (entityType === ENTITY_TYPE_CONTAINER_OBSERVED_DATA) {
-    const observed = stix as SDO.StixObservedData;
-    const from = observed.first_observed?.toISOString() ?? '-inf';
-    const to = observed.last_observed?.toISOString() ?? '+inf';
-    return `${from} - ${to}`;
-  }
-  if (entityType === ENTITY_TYPE_COURSE_OF_ACTION) {
-    return (stix as SDO.StixCourseOfAction).name;
-  }
-  if (entityType === ENTITY_TYPE_INCIDENT) {
-    return (stix as SDO.StixIncident).name;
-  }
-  if (entityType === ENTITY_TYPE_INDICATOR) {
-    return (stix as SDO.StixIndicator).name;
-  }
-  if (entityType === ENTITY_TYPE_INTRUSION_SET) {
-    return (stix as SDO.StixIntrusionSet).name;
-  }
-  if (entityType === ENTITY_TYPE_TOOL) {
-    return (stix as SDO.StixTool).name;
-  }
-  if (entityType === ENTITY_TYPE_VULNERABILITY) {
-    return (stix as SDO.StixVulnerability).name;
-  }
-  // endregion
-  // region meta entities
-  if (entityType === ENTITY_TYPE_MARKING_DEFINITION) {
-    return (stix as SMO.StixMarkingDefinition).name;
-  }
-  if (entityType === ENTITY_TYPE_LABEL) {
-    return (stix as SMO.StixLabel).value;
-  }
-  if (entityType === ENTITY_TYPE_EXTERNAL_REFERENCE) {
-    const externalRef = stix as SMO.StixExternalReference;
-    return `${externalRef.source_name}${externalRef.external_id ? ` (${externalRef.external_id})` : ''}`;
-  }
-  if (entityType === ENTITY_TYPE_KILL_CHAIN_PHASE) {
-    return (stix as SMO.StixKillChainPhase).kill_chain_name;
-  }
-  // endregion
-  // region Meta observable
-  if (entityType === ENTITY_WINDOWS_REGISTRY_VALUE_TYPE) {
-    const registry = stix as SCO.StixWindowsRegistryValueType;
-    return registry.name ?? registry.data ?? 'Unknown';
-  }
-  if (entityType === ENTITY_EMAIL_MIME_PART_TYPE) {
-    return (stix as SCO.StixEmailBodyMultipart).description;
-  }
-  // endregion
-  // region Observables
-  if (entityType === ENTITY_HASHED_OBSERVABLE_ARTIFACT) {
-    const artifact = stix as SCO.StixArtifact;
-    return hashValue(artifact) ?? artifact.payload_bin ?? artifact.url ?? 'Unknown';
-  }
-  if (entityType === ENTITY_AUTONOMOUS_SYSTEM) {
-    const autonomous = stix as SCO.StixAutonomousSystem;
-    return autonomous.name ?? autonomous.number ?? 'unknown';
-  }
-  if (entityType === ENTITY_BANK_ACCOUNT) {
-    const bankAccount = stix as SCO.StixBankAccount;
-    return bankAccount.iban ?? bankAccount.account_number ?? 'Unknown';
-  }
-  if (entityType === ENTITY_CRYPTOGRAPHIC_WALLET) {
-    return (stix as SCO.StixCryptocurrencyWallet).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_DIRECTORY) {
-    return (stix as SCO.StixDirectory).path ?? 'Unknown';
-  }
-  if (entityType === ENTITY_DOMAIN_NAME) {
-    return (stix as SCO.StixDomainName).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_EMAIL_ADDR) {
-    return (stix as SCO.StixEmailAddress).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_EMAIL_MESSAGE) {
-    const email = stix as SCO.StixEmailMessage;
-    return email.body ?? email.subject ?? 'Unknown';
-  }
-  if (entityType === ENTITY_HASHED_OBSERVABLE_STIX_FILE) {
-    const file = stix as SCO.StixFile;
-    return hashValue(file) ?? file.name ?? 'Unknown';
-  }
-  if (entityType === ENTITY_HOSTNAME) {
-    return (stix as SCO.StixHostname).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_IPV4_ADDR) {
-    return (stix as SCO.StixIPv4Address).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_IPV6_ADDR) {
-    return (stix as SCO.StixIPv6Address).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_MAC_ADDR) {
-    return (stix as SCO.StixMacAddress).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_MEDIA_CONTENT) {
-    const media = stix as SCO.StixMediaContent;
-    return media.content ?? media.title ?? media.url ?? 'Unknown';
-  }
-  if (entityType === ENTITY_MUTEX) {
-    return (stix as SCO.StixMutex).name ?? 'Unknown';
-  }
-  if (entityType === ENTITY_NETWORK_TRAFFIC) {
-    return String((stix as SCO.StixNetworkTraffic).dst_port ?? 'Unknown');
-  }
-  if (entityType === ENTITY_PROCESS) {
-    const process = stix as SCO.StixProcess;
-    return String(process.pid ?? process.command_line ?? 'Unknown');
-  }
-  if (entityType === ENTITY_SOFTWARE) {
-    return (stix as SCO.StixSoftware).name ?? 'Unknown';
-  }
-  if (entityType === ENTITY_TEXT) {
-    return (stix as SCO.StixText).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_PHONE_NUMBER) {
-    return (stix as SCO.StixPhoneNumber).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_PAYMENT_CARD) {
-    const paymentCard = stix as SCO.StixPaymentCard;
-    return paymentCard.card_number ?? paymentCard.holder_name ?? 'Unknown';
-  }
-  if (entityType === ENTITY_URL) {
-    return (stix as SCO.StixURL).value ?? 'Unknown';
-  }
-  if (entityType === ENTITY_USER_ACCOUNT) {
-    const userAccount = stix as SCO.StixUserAccount;
-    return userAccount.account_login ?? userAccount.user_id ?? 'Unknown';
-  }
-  if (entityType === ENTITY_WINDOWS_REGISTRY_KEY) {
-    return (stix as SCO.StixWindowsRegistryKey).key ?? 'Unknown';
-  }
-  if (entityType === ENTITY_HASHED_OBSERVABLE_X509_CERTIFICATE) {
-    const x509 = stix as SCO.StixX509Certificate;
-    return hashValue(x509) ?? x509.subject ?? x509.issuer ?? 'Unknown';
-  }
-  // endregion
-  throw UnsupportedError(`No representative extractor available for ${entityType}`);
+export const getStixRepresentativeConverters = (type: string) => {
+  return stixRepresentativeConverters.get(type);
 };
